requirements.txt
streamlit==1.23.0
torch>=2.0.0
transformers==4.31.0
timm==0.6.13
Pillow>=9.0.0
